Remove "IGI Airport, Delhi”
This has not chance since IGI isn’t in OSM and is an acronym. Official airport code is DEL,
so that’s not helpful either. We could boost certain words in the query
(for finding it in the name AND for matching category), such as airport or
station or bank, but that’s a future enhancement and possibly related to NLP.
Another example would be Hotel, should be part of the name AND category should be accommodation