Add confidence score test
This test lists a confidence score of 0.5 as an unexpected property. We
don't have > or < comparisons in our test suite, but I think the odds of
a well formed address query that we have good data for being given a
confidence score of 0.5, other than by hitting a deal-breaker (which
caps the score to exactly 0.5), is unlikely enough that this test will
work.