Add test cases for missing and incorrect postcodes
The first test is a case where the user has entered an incorrect, but
nearby zipcode for a document where we have the correct zipcode. We
should be able to figure things out anyways.

The second test has the user enter the correct zipcode, but the correct
document is missing zipcode info. We should also handle this just fine.