Mark coarse reverse test failing
Since we [changed the max radius for reverse](https://github.com/pelias/api/pull/549),
this test has been failing. The second part of the work to improve
reverse is to fall back to point-in-polygon lookup of admin areas if
there are no venues or POIs nearby, which has yet to be done.