Remove overspecified conditions from starbucks test
This test's purpose is to verify that with the query "starbucks 10010",
a venue named starbucks in the 10010 zipcode (in Mahattan, NYC), is
returned.

There's no way to say that a particular Starbucks should be returned, so
the housenumber and street name have been removed. Likewise, this isn't
a label test, so the label has been removed from the expected output.