Mark search focus point balance test failing
I was wondering why no tests alerted us of the focus point balance
issues in https://github.com/pelias/pelias/issues/849 and
https://github.com/pelias/openstreetmap/issues/507.

It turned out that the `priorityThresh` value in the relevant tests was
too generous.

This is now changed and we should be able to tell if/when progress is
made.