Remove autocomplete jitter test suite
This test suite has been 100% passing for a long time. Because it
contains a test for every single character of the input of an address,
it takes up quite a bit of screen real estate when evaluating tests.

The test in question has been moved to a single test case in the
autocomplete focus test suite

Connects https://github.com/pelias/pelias/issues/165