Add fallback test case
Fallback behavior is a core part of Pelias: rather than just returning
any results that match some of the search terms, we want Pelias to
return results with some knowledge of the hierarchy of elements in the
query.

This adds a test case confirming the behavior that, when given a query
for a street in a city, but the street doesn't exist Pelias returns the city.

The undesired behavior is that other streets with the same name in other
cities are returned.