Add tests for 2 letter country code handling in autocomplete
As of this writing, Pelias stores and indexes in Elasticsearch _only_ the 3 character [ISO 3166-1](https://en.wikipedia.org/wiki/List_of_ISO_3166_country_codes)
country code for every record.

For a select few countries, the 2 character code is not a prefix of the
3 character code (for example, Mexico is `MX`/`MEX`).

Most of Pelias can handle queries for either variant of the country code
just fine.

However, the autocomplete endpoint currently does _not_ return any
results if one of those few special two character codes is used.

While probably a fairly rare occurrence in actual usage, since they
would generally appear at the very end of a query, it's nice to ensure
these are handled properly.

This PR includes tests for 3 different autocomplete cases:
- query for a city
- two queries for addresses

Each cases is for a relevant non-prefix country code, and there are
tests for both the 2 and 3 letter codes, to ensure we're covering
things.

As of the creation of this PR, all the two letter country code query
variants fail when the second character in the country code is entered,
but pass otherwise once a reasonable number of characters has been
entered.

Related PRs that attempt to solve this problem:
https://github.com/pelias/schema/pull/472
https://github.com/pelias/schema/pull/470
https://github.com/pelias/schema/pull/469