Add leading-zero postalcode tests
This adds tests for a particular type of postalcode we've had trouble
with: those with a leading zero.

Tests are added to the existing 'search postalcode' suite, and a new
'autocomplete postalcode' sutie is also created.

We have https://github.com/pelias/schema/pull/475 in progress which we
hope can fix the current known issues.

Connects https://github.com/pelias/pelias/issues/898