Update 'Township' localities
It appears WOF recently added localities for several localadmins that
did not have a corresponding locality.

This really only makes a difference where the "official" name used for
the localadmin includes the word 'Township'. Generally the locality name
won't include 'Township'.

See the related change to _add_ township names to the localities in fdffa4fdd352.