Split Luxembourg country test
We now return Luxembourg the city instead of the country. This is
because of a data change that now means the two will be deduped
according to our normal logic (which prefers the more granular result).

However, we still want to test that the data for the country is returned
correctly, so the base test has been updated and a new country layer
specific test added.